+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Erl2ex is currently pre-alpha software. Expect significant backwards-incompatibl
 
 *   Much better reporting of parse and conversion errors
 *   Support for custom and remote types
+*   Support for receive with timeout
 
 ## v0.0.1 (2015-12-28)